Create DisplayTables.php
This is sctipt php to provide main information of your tables in database. You have to know DBName, DBHost, DBPass and DBUser to connect MySQL. So, If you have this information and you can upload the file per FTP, you can to get to know something of structure or values using the tables.